+++ R/wrcc_createDataDataframe.R
@@ -42,7 +42,7 @@ wrcc_createDataDataframe <- function(tbl, meta) {
     stop(paste0("Multiple monitor types found in 'meta' dataframe."))
   }
   
-  # Create monitorID the same way we did in airsis_createMetaDataframe()
+  # Create monitorID the same way we did in wrcc_createMetaDataframe()
   # Should only have a single instrumentID
   instrumentIDs <- sort(unique(meta$instrumentID))
   if ( length(instrumentIDs) > 1 ) {