+++ NEWS.md
@@ -14,9 +14,11 @@ from the US EPA, AirNow, AIRSIS, WRCC and others.
  * removed dependency on *RCurl* package
  * internal `WRCC` object with unitIDs is now a list of lists
  * more consistent error handling and logging during WRCC data processing
- * new `wrcc_load()` function to access pre-generated annual .RData files
  * updated `airnow_loadLatest()` function to access real-time, last 10 day files
  * new `airnow_loadDaily()` function to access daily-updated, last 45 day files
+ * new `wrcc_load()` function to access pre-generated annual .RData files
+ * updated `wrcc_loadLatest()` function to access real-time, last 10 day files
+ * new `wrcc_loadDaily()` function to access daily-updated, last 45 day files
 
 ### PWFSLSmoke 1.0.7